intelvestor-ml/requirements.txt
fastapi==0.118.0
uvicorn==0.22.0
yfinance==0.2.28
newsapi-python==0.2.7
transformers==4.31.0
langchain-google-genai==0.0.5
google-generativeai==0.3.2
pandas==2.0.3
numpy==1.25.2
scikit-learn==1.3.0
xgboost==1.7.6
tensorflow==2.14.0
statsmodels==0.14.0
prophet==1.1.5
numba==0.59.1
llvmlite==0.42.0
shap==0.44.0
python-dotenv==1.0.1
requests==2.32.3
torch==2.0.1
plotly==5.22.0
httpx==0.24.1
pydantic==2.11.9
textblob==0.17.1
nltk==3.8.1
scipy==1.11.1
